+++ packages/commercetools/theme/pages/CheckoutV2/Shipping.vue
@@ -8,7 +8,7 @@
     <VsfShippingProvider
       v-if="isShippingDetailsCompleted"
       :finished.sync="isShippingMethodCompleted"
-      @stepSubmit="handleStepSubmit"
+      @submit="handleStepSubmit"
     />
   </div>
 </template>
